+++ src/main/java/uk/gov/hmcts/reform/profilesync/service/ProfileSyncService.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.reform.profilesync.service;
 
 import java.util.List;
 
+import uk.gov.hmcts.reform.profilesync.advice.UserProfileSyncException;
 import uk.gov.hmcts.reform.profilesync.client.IdamClient;
-import uk.gov.hmcts.reform.profilesync.domain.UserProfileSyncException;
 
 public interface ProfileSyncService {
