+++ src/test/java/uk/gov/hmcts/reform/profilesync/util/UserProfileSyncJobSchedulerTest.java
@@ -7,9 +7,9 @@ import static org.mockito.Mockito.*;
 import java.time.LocalDateTime;
 
 import org.junit.Test;
+import uk.gov.hmcts.reform.profilesync.advice.UserProfileSyncException;
 import uk.gov.hmcts.reform.profilesync.domain.SyncJobAudit;
 import uk.gov.hmcts.reform.profilesync.domain.SyncJobConfig;
-import uk.gov.hmcts.reform.profilesync.domain.UserProfileSyncException;
 import uk.gov.hmcts.reform.profilesync.repository.SyncConfigRepository;
 import uk.gov.hmcts.reform.profilesync.repository.SyncJobRepository;
 import uk.gov.hmcts.reform.profilesync.service.ProfileSyncService;
